requirements.txt
undetected-chromedriver
selenium
nextcord[speed]
git+https://github.com/xtekky/gpt4free.git
git+https://github.com/QuartzWarrior/Quart-Nextcord.git
ujson
nextcord-ext-ipc
quart-rate-limiter
google-generativeai
ipython
google-api-python-client